Trait solana_bench_tps::bench_tps_client::BenchTpsClient

pub trait BenchTpsClient {

    // Required methods
    fn send_transaction(
        &self,
        transaction: Transaction
    ) -> Result<Signature, BenchTpsError>;
    fn send_batch(
        &self,
        transactions: Vec<Transaction>
    ) -> Result<(), BenchTpsError>;
    fn get_latest_blockhash(&self) -> Result<Hash, BenchTpsError>;
    fn get_latest_blockhash_with_commitment(
        &self,
        commitment_config: CommitmentConfig
    ) -> Result<(Hash, u64), BenchTpsError>;
    fn get_transaction_count(&self) -> Result<u64, BenchTpsError>;
    fn get_transaction_count_with_commitment(
        &self,
        commitment_config: CommitmentConfig
    ) -> Result<u64, BenchTpsError>;
    fn get_epoch_info(&self) -> Result<EpochInfo, BenchTpsError>;
    fn get_balance(&self, pubkey: &Pubkey) -> Result<u64, BenchTpsError>;
    fn get_balance_with_commitment(
        &self,
        pubkey: &Pubkey,
        commitment_config: CommitmentConfig
    ) -> Result<u64, BenchTpsError>;
    fn get_fee_for_message(
        &self,
        message: &Message
    ) -> Result<u64, BenchTpsError>;
    fn get_minimum_balance_for_rent_exemption(
        &self,
        data_len: usize
    ) -> Result<u64, BenchTpsError>;
    fn addr(&self) -> String;
    fn request_airdrop_with_blockhash(
        &self,
        pubkey: &Pubkey,
        lamports: u64,
        recent_blockhash: &Hash
    ) -> Result<Signature, BenchTpsError>;
    fn get_account(&self, pubkey: &Pubkey) -> Result<Account, BenchTpsError>;
    fn get_account_with_commitment(
        &self,
        pubkey: &Pubkey,
        commitment_config: CommitmentConfig
    ) -> Result<Account, BenchTpsError>;
    fn get_multiple_accounts(
        &self,
        pubkeys: &[Pubkey]
    ) -> Result<Vec<Option<Account>>, BenchTpsError>;

}

Required Methods

fn send_transaction( &self, transaction: Transaction ) -> Result<Signature, BenchTpsError>

Send a signed transaction without confirmation

fn send_batch( &self, transactions: Vec<Transaction> ) -> Result<(), BenchTpsError>

Send a batch of signed transactions without confirmation.

fn get_latest_blockhash(&self) -> Result<Hash, BenchTpsError>

Get latest blockhash

fn get_latest_blockhash_with_commitment( &self, commitment_config: CommitmentConfig ) -> Result<(Hash, u64), BenchTpsError>

Get latest blockhash and its last valid block height, using explicit commitment

fn get_transaction_count(&self) -> Result<u64, BenchTpsError>

Get transaction count

fn get_transaction_count_with_commitment( &self, commitment_config: CommitmentConfig ) -> Result<u64, BenchTpsError>

Get transaction count, using explicit commitment

fn get_epoch_info(&self) -> Result<EpochInfo, BenchTpsError>

Get epoch info

fn get_balance(&self, pubkey: &Pubkey) -> Result<u64, BenchTpsError>

Get account balance

fn get_balance_with_commitment( &self, pubkey: &Pubkey, commitment_config: CommitmentConfig ) -> Result<u64, BenchTpsError>

Get account balance, using explicit commitment

fn get_fee_for_message(&self, message: &Message) -> Result<u64, BenchTpsError>

Calculate the fee for a Message

fn get_minimum_balance_for_rent_exemption( &self, data_len: usize ) -> Result<u64, BenchTpsError>

Get the rent-exempt minimum for an account

fn addr(&self) -> String

Return the address of client

fn request_airdrop_with_blockhash( &self, pubkey: &Pubkey, lamports: u64, recent_blockhash: &Hash ) -> Result<Signature, BenchTpsError>

Request, submit, and confirm an airdrop transaction

fn get_account(&self, pubkey: &Pubkey) -> Result<Account, BenchTpsError>

Returns all information associated with the account of the provided pubkey

fn get_account_with_commitment( &self, pubkey: &Pubkey, commitment_config: CommitmentConfig ) -> Result<Account, BenchTpsError>

Returns all information associated with the account of the provided pubkey, using explicit commitment

fn get_multiple_accounts( &self, pubkeys: &[Pubkey] ) -> Result<Vec<Option<Account>>, BenchTpsError>